+++ crates/matrix-sdk-ui/CHANGELOG.md
@@ -13,6 +13,14 @@ All notable changes to this project will be documented in this file.
   non-left room filter did not take the new room stat into account.
   ([#4448](https://github.com/matrix-org/matrix-rust-sdk/pull/4448))
 
+### Features
+
+- [**breaking**] `Timeline::send_attachment()` now takes a type that implements
+  `Into<AttachmentSource>` instead of a type that implements `Into<PathBuf>`.
+  `AttachmentSource` allows to send an attachment either from a file, or with
+  the bytes and the filename of the attachment. Note that all types that
+  implement `Into<PathBuf>` also implement `Into<AttachmentSource>`.
+
 ## [0.9.0] - 2024-12-18
 
 ### Bug Fixes